+++ test/governance/extensions/GovernorCountingFractional.test.js
@@ -140,7 +140,7 @@ contract('GovernorCountingFractional', function (accounts) {
       const againstVotes = (new BN(voter2Weight)).mul(new BN(20)).div(new BN(100)); // 20 percent Against
       const abstainVotes = (new BN(voter2Weight)).sub(forVotes).sub(againstVotes);
 
-      const params = web3.eth.abi.encodeParameters(['uint128', 'uint128'], [forVotes, againstVotes]);
+      const params = web3.eth.abi.encodeParameters(['uint80', 'uint80'], [forVotes, againstVotes]);
       const tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', params, { from: voter2 });
 
       expectEvent(tx, 'VoteCastWithParams', { voter: voter2, weight: voter2Weight, params });
@@ -183,7 +183,7 @@ contract('GovernorCountingFractional', function (accounts) {
       const againstVotes = new BN(0);
       const abstainVotes = new BN(voter2Weight);
 
-      const params = web3.eth.abi.encodeParameters(['uint128', 'uint128'], [forVotes, againstVotes]);
+      const params = web3.eth.abi.encodeParameters(['uint80', 'uint80'], [forVotes, againstVotes]);
       const tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', params, { from: voter2 });
 
       expectEvent(tx, 'VoteCastWithParams', { voter: voter2, weight: voter2Weight, params });
@@ -226,7 +226,7 @@ contract('GovernorCountingFractional', function (accounts) {
       const againstVotes = new BN(0);
       const abstainVotes = new BN(0);
 
-      const params = web3.eth.abi.encodeParameters(['uint128', 'uint128'], [forVotes, againstVotes]);
+      const params = web3.eth.abi.encodeParameters(['uint80', 'uint80'], [forVotes, againstVotes]);
       const tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', params, { from: voter2 });
 
       expectEvent(tx, 'VoteCastWithParams', { voter: voter2, weight: voter2Weight, params });
@@ -269,7 +269,7 @@ contract('GovernorCountingFractional', function (accounts) {
       const againstVotes = new BN(voter2Weight);
       const abstainVotes = new BN(0);
 
-      const params = web3.eth.abi.encodeParameters(['uint128', 'uint128'], [forVotes, againstVotes]);
+      const params = web3.eth.abi.encodeParameters(['uint80', 'uint80'], [forVotes, againstVotes]);
       const tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', params, { from: voter2 });
 
       expectEvent(tx, 'VoteCastWithParams', { voter: voter2, weight: voter2Weight, params });
@@ -320,7 +320,7 @@ contract('GovernorCountingFractional', function (accounts) {
       const voter3AbstainVotes = (new BN(voter3Weight)).sub(voter3ForVotes).sub(voter3AgainstVotes);
 
       // voter 2 casts votes
-      const voter2Params = web3.eth.abi.encodeParameters(['uint128', 'uint128'], [voter2ForVotes, voter2AgainstVotes]);
+      const voter2Params = web3.eth.abi.encodeParameters(['uint80', 'uint80'], [voter2ForVotes, voter2AgainstVotes]);
       const voter2Tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', voter2Params, { from: voter2 });
       expectEvent(
         voter2Tx,
@@ -333,7 +333,7 @@ contract('GovernorCountingFractional', function (accounts) {
       expect(votes.abstainVotes).to.be.bignumber.equal(voter2AbstainVotes);
 
       // voter 2 casts votes
-      const voter3Params = web3.eth.abi.encodeParameters(['uint128', 'uint128'], [voter3ForVotes, voter3AgainstVotes]);
+      const voter3Params = web3.eth.abi.encodeParameters(['uint80', 'uint80'], [voter3ForVotes, voter3AgainstVotes]);
 const voter3Tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', voter3Params, { from: voter3 });
       expectEvent(
         voter3Tx,
@@ -383,7 +383,7 @@ const voter3Tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', vot
       const againstVotes = (new BN(voter2Weight)).mul(new BN(1)).div(new BN(100)); // 1%
       const abstainVotes = (new BN(voter2Weight)).sub(forVotes).sub(againstVotes);
 
-      const params = web3.eth.abi.encodeParameters(['uint128', 'uint128'], [forVotes, againstVotes]);
+      const params = web3.eth.abi.encodeParameters(['uint80', 'uint80'], [forVotes, againstVotes]);
       const tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', params, { from: voter2 });
 
       expectEvent(tx, 'VoteCastWithParams', { voter: voter2, weight: voter2Weight, params });
@@ -439,7 +439,7 @@ const voter3Tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', vot
       const againstVotes = (new BN(voter2Weight)).mul(new BN(90)).div(new BN(100)); // 90%
       const abstainVotes = (new BN(voter2Weight)).sub(forVotes).sub(againstVotes);
 
-      const params = web3.eth.abi.encodeParameters(['uint128', 'uint128'], [forVotes, againstVotes]);
+      const params = web3.eth.abi.encodeParameters(['uint80', 'uint80'], [forVotes, againstVotes]);
       const tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', params, { from: voter2 });
 
       expectEvent(tx, 'VoteCastWithParams', { voter: voter2, weight: voter2Weight, params });
@@ -500,7 +500,7 @@ const voter3Tx = await this.mock.castVoteWithReasonAndParams(this.id, 0, '', vot
         "test assumption not met"
       )
 
-      const params = web3.eth.abi.encodeParameters(['uint128', 'uint128'], [forVotes, againstVotes]);
+      const params = web3.eth.abi.encodeParameters(['uint80', 'uint80'], [forVotes, againstVotes]);
       await expectRevert(
         this.mock.castVoteWithReasonAndParams(this.id, 0, '', params, { from: voter2 }),
         'GovernorCountingFractional: Invalid Weight'
